refactored landing and styling into components and partials
