+++ src/components/common/Footer.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <footer>
+      <p>things</p>
+      <p>things</p>
+      <p>and morethings</p>
+    </footer>
+  )
+}
+export default Footer
