+++ src/components/landing/Feature.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+const Feature = () => {
+  return (
+    <section className='flex-container s16' id='s3'>
+      <h3>Results</h3>
+      <div className=''>
+        <div>&quot;Quote, testimonal from client&quot;</div>
+
+        <div className=''>Image of before and after</div>
+
+        {/* build login page for rhyse, user authentication */}
+        {/* clickthrough to success stories, build small-scale CMS for ryhse to self-publish */}
+      </div>
+    </section>
+  )
+}
+export default Feature
