+++ src/components/landing/Services.js
@@ -1,0 +1,58 @@
+import React from 'react'
+
+const Services = () => {
+  return (
+    <section className='' id='services'>
+      <h3>Personal Training</h3>
+      <h3>Massage & Soft Tissue Therapy</h3>
+
+      <div className='gridbox'>
+        <li className='gridcell'>
+          <img
+            className='grid-icon'
+            src='https://cdn4.iconfinder.com/data/icons/oakcons-2/16/Image-512.png'
+          ></img>
+          <h3>Service 1</h3>
+          <p>
+            Blah blah blah, blah, blah-blah-blah. Blah blah blah, blah,
+            blah-blah-blah.
+          </p>
+        </li>
+        <li className='gridcell'>
+          <img
+            className='grid-icon'
+            src='https://cdn4.iconfinder.com/data/icons/oakcons-2/16/Image-512.png'
+          ></img>
+          <h3>Service 2</h3>
+          <p>
+            Blah blah blah, blah, blah-blah-blah. Blah blah blah, blah,
+            blah-blah-blah.
+          </p>
+        </li>
+        <li className='gridcell'>
+          <img
+            className='grid-icon'
+            src='https://cdn4.iconfinder.com/data/icons/oakcons-2/16/Image-512.png'
+          ></img>
+          <h3>Service 3</h3>
+          <p>
+            Blah blah blah, blah, blah-blah-blah.Blah blah blah, blah,
+            blah-blah-blah.
+          </p>
+        </li>
+        <li className='gridcell'>
+          <img
+            className='grid-icon'
+            src='https://cdn4.iconfinder.com/data/icons/oakcons-2/16/Image-512.png'
+          ></img>
+          <h3>Service 4</h3>
+          <p>
+            Blah blah blah, blah, blah-blah-blah. Blah blah blah, blah,
+            blah-blah-blah.
+          </p>
+        </li>
+      </div>
+    </section>
+  )
+}
+export default Services
