+++ src/styles/contact_form.scss
@@ -10,7 +10,7 @@
   margin: 0 auto;
 
   input, textarea {
-    border: 1px solid grey;
+    border: 1px solid silver;
     border-radius: 3px;
     width: 600px;
     height: 35px;
@@ -24,6 +24,7 @@
     margin: auto;
   }
   span > input {
+    margin-right: -4px;
     width: 47%;
     display: flex;
     justify-content: space-between;
